Fixing more of Julio.

+++ MansionMashers/MansionMashers/HeaderFiles/Camera.h
@@ -1,6 +1,6 @@
 #ifndef FOX_CAMERA
 #define FOX_CAMERA
 
-void SF_SetCamera(float playerX, int DistUntilCamMove, int CameraMoveX);
+void SF_SetCamera(float playerX, int DistUntilCamMove, float CameraMoveX);
 
 #endif
